+++ plzdrawing/src/components/common/text/Txt.tsx
@@ -11,6 +11,7 @@ interface TextProps extends RNTextProps {
   children: React.ReactNode;
   style?: StyleProp<TextStyle>;
   weight?: "bold" | "light";
+  align?: "center" | "left" | "right";
   color?: keyof typeof colors.colors; // Colors 객체의 키를 타입으로
   size?: number; // 텍스트 크기
   lineheight?: number; // 텍스트 높이
@@ -22,7 +23,7 @@ export default function Txt({
   weight = "light",
   color = "black", // 기본 색상
   size = 14, // 기본 크기
-  lineheight = size,
+  lineheight,
   ...props
 }: TextProps) {
   return (
@@ -35,8 +36,8 @@ export default function Txt({
           fontSize: size,
           color: colors.colors[color], // Colors 객체에서 색상 가져오기
           fontFamily: weight === "bold" ? "Ssurround" : "SsurroundAir",
-          textAlign: "center",
-          lineHeight: lineheight,
+          textAlign: props.align,
+          lineHeight: lineheight ? lineheight : size * 1.2,
         },
       ]}
     >
